[#931] Started restructirng to dedicated _class08_get_data.py and _class08_concatenate.py modules
