+++ tofu/data/_class08_Diagnostic.py
@@ -13,6 +13,8 @@ import numpy as np
 from ._class07_Camera import Camera as Previous
 from . import _class8_check as _check
 from . import _class8_compute as _compute
+from . import _class08_get_data as _get_data
+from . import _class08_concatenate as _concatenate
 from . import _class8_move as _move
 from . import _class8_los_data as _los_data
 from . import _class8_equivalent_apertures as _equivalent_apertures
@@ -172,7 +174,7 @@ class Diagnostic(Previous):
             'res'
 
         """
-        return _compute._get_data(
+        return _get_data._get_data(
             coll=self,
             key=key,
             key_cam=key_cam,
@@ -194,7 +196,7 @@ class Diagnostic(Previous):
 
 
         """
-        return _compute._concatenate_data(
+        return _concatenate._concatenate_data(
             coll=self,
             key=key,
             key_data=key_data,
